refactor: move files to corresponding folders
