+++ pyproject.toml
@@ -35,6 +35,7 @@ pydantic = "^1.10.1"
 ratelimit = "^2.2.1"
 loguru = "^0.6.0"
 validators = "^0.20.0"
+asyncio = "^3.4.3"
 
 [tool.poetry.dev-dependencies]
 pylint = "^2.15.0"
