+++ shikithon/enums/comment.py
@@ -1,5 +1,5 @@
 """Enums for /api/comments."""
-from shikithon.enums.enhanced_enum import EnhancedEnum
+from .enhanced_enum import EnhancedEnum
 
 
 class CommentableType(EnhancedEnum):
