+++ shikithon/enums/favorite.py
@@ -1,5 +1,5 @@
 """Enums for /api/favorites."""
-from shikithon.enums.enhanced_enum import EnhancedEnum
+from .enhanced_enum import EnhancedEnum
 
 
 class FavoriteLinkedType(EnhancedEnum):
