+++ shikithon/enums/ranobe.py
@@ -1,5 +1,5 @@
 """Enums for /api/ranobe."""
-from shikithon.enums.enhanced_enum import EnhancedEnum
+from .enhanced_enum import EnhancedEnum
 
 
 class RanobeOrder(EnhancedEnum):
