+++ shikithon/exceptions/__init__.py
@@ -1,0 +1,14 @@
+"""Exceptions for shikithon API class.
+
+Used for raising on validation of config/app data.
+"""
+
+from .access_token_exception import AccessTokenException
+from .missing_app_variable import MissingAppVariable
+from .missing_auth_code import MissingAuthCode
+from .missing_config_data import MissingConfigData
+
+__all__ = [
+    'AccessTokenException', 'MissingAuthCode', 'MissingAppVariable',
+    'MissingConfigData'
+]
