+++ shikithon/models/message.py
@@ -4,8 +4,8 @@ from typing import Optional
 
 from pydantic import BaseModel, Field
 
-from shikithon.models.linked_topic import LinkedTopic
-from shikithon.models.user import User
+from .linked_topic import LinkedTopic
+from .user import User
 
 
 class Message(BaseModel):
