+++ shikithon/models/people_works.py
@@ -3,8 +3,8 @@ from typing import Optional
 
 from pydantic import BaseModel
 
-from shikithon.models.anime import Anime
-from shikithon.models.manga import Manga
+from .anime import Anime
+from .manga import Manga
 
 
 class PeopleWorks(BaseModel):
