+++ shikithon/models/user_list.py
@@ -4,9 +4,9 @@ from typing import Optional
 
 from pydantic import BaseModel
 
-from shikithon.models.anime import Anime
-from shikithon.models.manga import Manga
-from shikithon.models.user import User
+from .anime import Anime
+from .manga import Manga
+from .user import User
 
 
 class UserList(BaseModel):
